KTOR-2763 Fix warnings in ktor-io

+++ ktor-io/jvm/src/io/ktor/utils/io/internal/Strings.kt
@@ -115,7 +115,7 @@ private fun ByteBuffer.decodeASCII3_array(out: CharArray, offset: Int, length: I
             val b = array[srcPos]
             if (b < 0) break
 
-            out[pos] = b.toChar()
+            out[pos] = b.toInt().toChar()
 
             pos++
             srcPos++
@@ -144,7 +144,7 @@ private fun ByteBuffer.decodeASCII3_buffer(out: CharArray, offset: Int, length: 
                 pushBack = true
                 break
             }
-            out[pos] = b.toChar()
+            out[pos] = b.toInt().toChar()
             pos++
         }
     }
@@ -173,7 +173,7 @@ private inline fun ByteBuffer.decodeASCII3_array(
             val b = array[srcPos]
             if (b < 0) break
 
-            val ch = b.toChar()
+            val ch = b.toInt().toChar()
             if (!predicate(ch)) {
                 position(srcPos - arrayOffset())
                 return decodeUtf8Result(pos - offset, -1)
@@ -214,7 +214,7 @@ private inline fun ByteBuffer.decodeASCII3_buffer(
                 break
             }
 
-            val ch = b.toChar()
+            val ch = b.toInt().toChar()
             if (!predicate(ch)) {
                 pushBack = true
                 predicateFailed = true
